refactor: tidy up paths and button docs
