+++ docs/components/LinkRenderer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Link } from '../../src/components/link';
+import { Link } from '../../src';
 
 export interface LinkRendererProps {
     readonly children?: React.ReactNode;
