+++ package.json
@@ -68,6 +68,7 @@
         "jest": "26.6.3",
         "jest-styled-components": "7.0.4",
         "lint-staged": "11.0.0",
+        "process": "^0.11.10",
         "react": "17.0.2",
         "react-docgen-typescript": "1.22.0",
         "react-dom": "17.0.2",
