+++ webpack.config.js
@@ -46,6 +46,9 @@ module.exports = {
         new webpack.DefinePlugin({
             'process.env.NODE_ENV': JSON.stringify(env)
         }),
+        new webpack.ProvidePlugin({
+            process: 'process/browser'
+        }),
         new CircularDependencyPlugin({
             exclude: /node_modules/,
             failOnError: true,
